Add:検索バー選択時に表示する、タグ検索ボタンのコンポーネントを追加しました
