--- miniApp/frontend/app/search/page.module.css
@@ -1,70 +1,0 @@
-
-.container {
-    padding: 24px 20px calc(env(safe-area-inset-bottom) + var(--bottom-nav-height, 65px) + 90px) 20px; /* 下部に固定する要素が被らないようにpadding-bottomを動的に計算 */
-    background-color: #ffffff;
-    font-family: sans-serif;
-    color: #333333;
-    min-height: 100vh;
-}
-
-.pageTitle {
-    font-size: var(--title-size);
-    font-weight: bold;
-    margin-bottom: 24px;
-    margin-top: 0;
-}
-
-.section {
-    margin-bottom: 32px;
-}
-
-.sectionTitle {
-    font-size: var(--subtitle-size);
-    font-weight: bold;
-    margin-bottom: 16px;
-    margin-top: 0;
-}
-
-.buttonGrid {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 12px;
-}
-
-/* MUI Buttonのベーススタイル上書き */
-.customButton {
-    width: 100%;
-    height: 52px;
-    border-radius: 8px !important;
-    font-size: var(--subtitle-size) !important;
-    font-weight: bold !important;
-    color: #ffffff !important;
-    letter-spacing: 1px !important;
-    /* MUI特有の英語大文字化設定を無効化 */
-    text-transform: none !important; 
-}
-
-/* 濃い緑のボタン */
-.buttonDark {
-    background-color: var(--primary-color) !important;
-}
-
-/* 薄い緑のボタン */
-.buttonLight {
-    background-color: var(--secondary-color) !important;
-}
-
-/* 検索バー */
-.searchTextField {
-    position: fixed;
-    /* layout.tsx のナビゲーションバーの safe-area を考慮した位置に合わせる */
-    bottom: calc(env(safe-area-inset-bottom) + var(--bottom-nav-height, 65px));
-    left: 0;
-    width: 100%;
-    padding: 10px 20px 10px 20px;
-    background-color: var(--main-bg-color);
-    z-index: 1000;
-    /* スクロール時の境界を分かりやすくするための境界線と影 */
-    border-top: 1px solid #eeeeee;
-    box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.05);
-}
