+++ miniApp/frontend/components/organisms/tagSearchButtons/TagSearchButtons.module.css
@@ -1,0 +1,39 @@
+.section {
+    margin-bottom: 32px;
+}
+
+.sectionTitle {
+    font-size: var(--subtitle-size);
+    font-weight: bold;
+    margin-bottom: 16px;
+    margin-top: 0;
+}
+
+.buttonGrid {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 12px;
+}
+
+/* MUI Buttonのベーススタイル上書き */
+.customButton {
+    width: 100%;
+    height: 52px;
+    border-radius: 8px !important;
+    font-size: var(--subtitle-size) !important;
+    font-weight: bold !important;
+    color: #ffffff !important;
+    letter-spacing: 1px !important;
+    /* MUI特有の英語大文字化設定を無効化 */
+    text-transform: none !important; 
+}
+
+/* 濃い緑のボタン */
+.buttonDark {
+    background-color: var(--primary-color) !important;
+}
+
+/* 薄い緑のボタン */
+.buttonLight {
+    background-color: var(--secondary-color) !important;
+}
